setupMounts throws an error if run on a stopped VM
The 9p mount already exists in this case.  This is reverting back to behavior before
https://github.com/zchee/docker-machine-driver-xhyve/pull/161/files

This now logs a warning, but this is a not fatal error on a recreate.  I'm not sure there's
a cleaner way since Start() has no context of if it is called on new VM or a recently stopped VM.