Do not use (and overwrite) `bootlocal.sh` to setup mounts anymore
Instead, simply prepare and execute the `mount` commands directly with `driver.RunSSHCommandFromDriver()` when `Start()` is called.

When reading https://github.com/boot2docker/boot2docker/blob/master/doc/FAQ.md#local-customisation-with-persistent-partition it appears that the intent of the `bootlocal.sh` script is to be a script created and managed by the user.

Which means that this driver, by overwriting the `bootlocal.sh` script, is going against what has been defined as the way of using this file.

This commit fixes this issue.